Speed improvement
Speed of the function is improved somewhat by loading only the needed data from the csv file.
This uses opts.SelectedVariableNames to specify which columns to get.
If only 2 columns are needed (timestamps and skin conductance) the function takes ~half the time

Columns that are not needed can be omitted by using
cfg.columnname.acc_x = 'ignore';
to call shimmer2matlab