adjusted data structs
Splitting up the data in the separate processing parts: segmentation, artifact correction, and deconvolution.

Zscored conductance is no longer stored for the segmented and artifact corrected data as this is simply the normalize() version of the conductance data. Zscore is calculated when needed for deconvolution.

Dat is downsampled to 4Hz before deconvolution to speed up processing (2x as fast) and minimize filesize (~30x smaller)